projects/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlmodel>=0.0.16
passlib>=1.7.4
python-jose[cryptography]>=3.3
apscheduler>=3.10
feedparser>=6.0
httpx>=0.25
python-dateutil>=2.9
pydantic>=2.5
tomli>=2.0
python-dotenv>=1.0
